@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "start": "SET NODE_ENV=production && node server/server.js",
-    "develop": "npm-run-all --parallel serverdebug:open *:watch",
-    "serverdebug:open": "SET NODE_ENV=development && nodemon server/server.js --watch ./server",
+    "develop": "npm-run-all --parallel dev-server *:watch",
+    "dev-server": "SET NODE_ENV=development && nodemon server/server.js --watch ./server",
     "prebuild": "npm-run-all --parallel test lint",
     "build": "rimraf public && webpack --config ./server/settings/webpack.config.prod",
     "preadd": "npm run build",
@@ -20,6 +20,8 @@
   "author": "Idan Avrahami",
   "license": "MIT",
   "dependencies": {
+    "@material-ui/core": "^4.9.5",
+    "@material-ui/icons": "^4.9.1",
     "axios": "^0.19.1",
     "body-parser": "^1.19.0",
     "compression": "^1.7.4",
@@ -27,6 +29,7 @@
     "helmet": "^3.21.2",
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.15",
+    "material-table": "^1.57.2",
     "mongodb": "^3.5.1",
     "mongoose": "^5.8.9",
     "prop-types": "^15.7.2",
@@ -59,22 +62,22 @@
     "eslint-watch": "^6.0.1",
     "html-webpack-plugin": "^3.2.0",
     "ignore-styles": "^5.0.1",
-    "jsdom": "^16.0.1",
     "json-loader": "^0.5.7",
     "mini-css-extract-plugin": "^0.9.0",
     "mocha": "^7.0.0",
+    "node-sass": "^4.13.1",
     "nodemon": "^2.0.2",
     "npm-run-all": "^4.1.5",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "sass-loader": "^8.0.2",
     "style-loader": "^1.1.3",
     "terser-webpack-plugin": "^2.3.2",
     "url-loader": "^3.0.0",
-    "webpack": "^4.41.5",
+    "webpack": "^4.42.0",
     "webpack-cli": "^3.3.10",
     "webpack-dev-middleware": "^3.7.2",
     "webpack-hot-middleware": "^2.25.0",
-    "webpack-md5-hash": "^0.0.6",
-    "yarn": "^1.21.1"
+    "webpack-md5-hash": "^0.0.6"
   },
   "babel": {
     "presets": [
