@@ -1,31 +1,21 @@
 export const creatingUserMsg = 'creatingUserMsg';
 import axios from 'axios';
-// import { LoginFunc } from '../login/login.action';
 
-export function createUserFunc(data) {
+export function createUserAction(data) {
     return (dispatch) => {
-
         if (data.error)
-            dispatch({ type: creatingUserMsg, msg: data.error });
-        else {
-            dispatch({ type: creatingUserMsg, msg: 'Creating user..' });
-
-            axios.post('/api/users', data)
-                .then((res) => {
-                    if (res.status === 201) {
-                        dispatch({ type: creatingUserMsg, msg: '' });
-                    } else
-                        dispatch({ type: creatingUserMsg, msg: res.data });
-
-                })
-                .catch((err) => {
-                    const errMsg = err.response.data;
-                    console.log(errMsg);
-                    if (errMsg && typeof(errMsg) === 'string')
-                        dispatch({ type: creatingUserMsg, msg: errMsg });
-                    else
-                        dispatch({ type: creatingUserMsg, msg: 'User creation failed' });
-                });
-        }
+            return new Promise((resolve) => resolve(dispatch({ type: creatingUserMsg, msg: data.error })));
+        return axios.post('/api/users', data)
+            .then((res) => {
+                if (res.status === 201)
+                    return dispatch({ type: creatingUserMsg, msg: '' });
+                return dispatch({ type: creatingUserMsg, msg: res.data });
+            })
+            .catch((err) => {
+                const errMsg = err.response.data;
+                if (errMsg && typeof (errMsg) === 'string')
+                    return dispatch({ type: creatingUserMsg, msg: errMsg });
+                return dispatch({ type: creatingUserMsg, msg: 'User creation failed' });
+            });
     };
 }
